Move Hibernate DDL script generation tests to non-Panache hibernate-orm / hibernate-reactive tests
These tests have no relationship whatsoever with Panache, and them being
prod-mode tests make them very problematic for reliance on dev services.
