+++ extensions/hibernate-reactive/deployment/src/test/resources/application-generate-script.properties
@@ -1,6 +1,10 @@
-quarkus.datasource.db-kind=h2
-quarkus.datasource.jdbc.url=jdbc:h2:mem:test
+quarkus.datasource.db-kind=postgresql
+quarkus.datasource.username=hibernate_orm_test
+quarkus.datasource.password=hibernate_orm_test
+quarkus.datasource.reactive=true
+quarkus.datasource.reactive.url=${postgres.reactive.url}
 
+# Hibernate config
 #quarkus.hibernate-orm.log.sql=true
 quarkus.hibernate-orm.schema-management.strategy=drop-and-create
