--- integration-tests/hibernate-orm-panache/src/main/resources/ddlgeneration.properties
@@ -1,12 +1,0 @@
-quarkus.datasource.db-kind=h2
-quarkus.datasource.jdbc.url=jdbc:h2:tcp://localhost/mem:test
-quarkus.datasource.jdbc.max-size=8
-
-quarkus.hibernate-orm.schema-management.strategy=drop-and-create
-
-quarkus.hibernate-orm.scripts.generation=drop-and-create
-quarkus.hibernate-orm.scripts.generation.create-target=create.sql
-quarkus.hibernate-orm.scripts.generation.drop-target=drop.sql
-
-quarkus.hibernate-orm.statistics=true
-quarkus.hibernate-orm.metrics.enabled=true
