--- integration-tests/hibernate-reactive-panache-kotlin/src/main/resources/ddlgeneration.properties
@@ -1,8 +1,0 @@
-quarkus.datasource.db-kind=postgresql
-quarkus.datasource.username=hibernate_orm_test
-quarkus.datasource.password=hibernate_orm_test
-quarkus.datasource.reactive.url=${postgres.reactive.url}
-
-quarkus.hibernate-orm.scripts.generation=drop-and-create
-quarkus.hibernate-orm.scripts.generation.create-target=create.sql
-quarkus.hibernate-orm.scripts.generation.drop-target=drop.sql
